requirements.txt
# Telegram Bot API
python-telegram-bot>=22.0,<23.0

# База данных PostgreSQL
psycopg2-binary>=2.9.9

# S3-совместимое облачное хранилище (AWS S3, MinIO, Yandex Cloud и т.д.)
boto3>=1.34.0

# Загрузка переменных окружения из .env файла
python-dotenv>=1.0.0

# Генерация PDF (для инструкций)
reportlab>=4.0.0

# Транскрибация голосовых (нужен ffmpeg в системе)
openai-whisper>=20231117
faster-whisper>=1.0.0

# Личный кабинет (веб)
Flask>=3.0.0
gunicorn>=21.0.0
passlib[bcrypt]>=1.7.4

# LLM для обработки транскрипта (биография)
openai>=1.0.0

# Диаризация (разделение по спикерам) — опционально
# pip install librosa scikit-learn
librosa>=0.10.0
scikit-learn>=1.0.0
